refactor(runtime): move manual history-compaction family into AiSdkCompaction
Move compactHistory, buildManualHistoryCompactPolicy, hasHistoryCompactWriter,
and hasHistoryCompactCheckpointWriter from AiSdkBackend into AiSdkCompaction.
historyCompactAbortController now lives on the collaborator (it is
manual-compaction-operation-scoped); AiSdkBackend.stop delegates to a new
compaction.abortHistoryCompact() hook, so the stop-then-new-send-overlap
behavior is unchanged. compactHistory becomes a public one-line delegate on
AiSdkBackend (the AgentBackend interface method) that threads
this.priorRequestShape?.requestShapeHash into the collaborator.

hasHistoryCompactWriter / hasHistoryCompactCheckpointWriter stay public on
the collaborator because buildPriorMessages (which stays in AiSdkBackend) also
calls them; those two call sites now go through this.compaction.*. The moved
methods' internal writeHistoryCompact* calls are now intra-collaborator
(this.*) and thread the requestShapeHashBefore param directly.

Behavior-neutral pure move: runtime build + 2073 tests pass (0 fail, 7
skipped), parity with the baseline. Part of #1084.
