+++ packages/runtime/src/ai-sdk-backend.ts
@@ -913,7 +913,6 @@ export class AiSdkBackend implements AgentBackend {
 
   private aborted = false;
   private abortController: AbortController | null = null;
-  private historyCompactAbortController: AbortController | null = null;
   private currentTurnId: string | null = null;
   private stopAfterStepRequested = false;
   private currentInvocationId: string | null = null;
@@ -1044,139 +1043,7 @@ export class AiSdkBackend implements AgentBackend {
   // --------------------------------------------------------------------------
 
   async compactHistory(input: BackendCompactHistoryInput): Promise<BackendCompactHistoryResult> {
-    const historyCompactAbortController = new AbortController();
-    this.historyCompactAbortController = historyCompactAbortController;
-    try {
-      const runtimeContext = input.runtimeContext.filter((event) => event.turnId !== input.turnId);
-      const policy = this.buildManualHistoryCompactPolicy(runtimeContext);
-      if (!policy) return {};
-
-      const contextBudget = policy;
-      const budgeted = applyRuntimeEventContextBudget(runtimeContext, contextBudget, {
-        historyCompactProtocol: this.hasHistoryCompactCheckpointWriter()
-          ? 'checkpoint_v2'
-          : 'legacy_v1',
-      });
-      let contextBudgetDiagnostic = budgeted?.diagnostic;
-
-      if (
-        budgeted?.historyCompactBlocks?.length &&
-        contextBudget.historyCompact?.mode === 'read_write' &&
-        this.hasHistoryCompactWriter()
-      ) {
-        const loadedBlockIds = new Set(
-          (contextBudget.historyCompact.blocks ?? []).map((block) => block.blockId),
-        );
-        const draftBlocks = budgeted.historyCompactBlocks.filter(
-          (block) => !loadedBlockIds.has(block.blockId),
-        );
-        if (draftBlocks.length > 0) {
-          if (this.input.summarizeHistoryCompact && this.input.recordHistoryCompactCheckpoint) {
-            let writeContextBudget = contextBudget;
-            try {
-              const checkpoint = await Promise.resolve(this.input.loadHistoryCompactCheckpoint?.());
-              if (checkpoint) {
-                writeContextBudget = {
-                  ...contextBudget,
-                  historyCompact: { ...contextBudget.historyCompact!, checkpoint },
-                };
-              }
-            } catch {
-              // A missing previous checkpoint only loses rolling reuse; the current fold remains safe to summarize.
-            }
-            const writePatch = await this.compaction.writeHistoryCompactCheckpoint({
-              turnId: input.turnId,
-              contextBudget: writeContextBudget,
-              priorRuntimeContext: runtimeContext,
-              draftBlock: draftBlocks[0]!,
-              abortSignal: historyCompactAbortController.signal,
-              requestShapeHashBefore: this.priorRequestShape?.requestShapeHash,
-            });
-            if (historyCompactAbortController.signal.aborted) return {};
-            contextBudgetDiagnostic = mergeContextBudgetDiagnostic(
-              contextBudgetDiagnostic ??
-                buildContextBudgetDiagnosticShell(runtimeContext, budgeted.events, contextBudget),
-              writePatch.diagnosticPatch,
-            );
-          } else {
-            const writePatch = await this.compaction.writeHistoryCompactBlocks({
-              turnId: input.turnId,
-              contextBudget,
-              priorRuntimeContext: runtimeContext,
-              draftBlocks,
-              abortSignal: historyCompactAbortController.signal,
-              requestShapeHashBefore: this.priorRequestShape?.requestShapeHash,
-            });
-            if (historyCompactAbortController.signal.aborted) return {};
-            if (writePatch.replacementBlocks.length === 0) {
-              contextBudgetDiagnostic = buildContextBudgetDiagnosticShell(
-                runtimeContext,
-                runtimeContext,
-                contextBudget,
-              );
-            }
-            contextBudgetDiagnostic = mergeContextBudgetDiagnostic(
-              contextBudgetDiagnostic ??
-                buildContextBudgetDiagnosticShell(runtimeContext, budgeted.events, contextBudget),
-              writePatch.diagnosticPatch,
-            );
-          }
-        }
-      }
-
-      return contextBudgetDiagnostic ? { contextBudget: contextBudgetDiagnostic } : {};
-    } finally {
-      if (this.historyCompactAbortController === historyCompactAbortController) {
-        this.historyCompactAbortController = null;
-      }
-    }
-  }
-
-  private buildManualHistoryCompactPolicy(
-    runtimeContext: readonly RuntimeEvent[],
-  ): ContextBudgetPolicy | undefined {
-    if (runtimeContext.length === 0 || !this.input.contextBudget || !this.hasHistoryCompactWriter())
-      return undefined;
-    const base = this.input.contextBudget;
-    const charsPerToken = base.charsPerToken ?? 4;
-    const estimatedTokens = Math.max(1, estimateRuntimeEventsTokens(runtimeContext, charsPerToken));
-    const current = base.historyCompact;
-    const currentWithoutBlocks = { ...current };
-    delete currentWithoutBlocks.blocks;
-    delete currentWithoutBlocks.checkpoint;
-    const maxHistoryEstimatedTokens =
-      base.maxHistoryEstimatedTokens ?? Math.max(estimatedTokens, 32_000);
-    return {
-      name: base.name ?? 'manual-history-compact',
-      ...(base.charsPerToken !== undefined ? { charsPerToken: base.charsPerToken } : {}),
-      maxHistoryEstimatedTokens,
-      minRecentTurns: current?.minRecentTurns ?? base.minRecentTurns ?? 1,
-      historyCompact: {
-        ...currentWithoutBlocks,
-        enabled: true,
-        mode: 'read_write',
-        highWaterRatio: 0.000001,
-        targetRatio: current?.targetRatio ?? 0.2,
-        tailEstimatedTokens: 1,
-        minRecentTurns: current?.minRecentTurns ?? base.minRecentTurns ?? 1,
-        maxBlocks: current?.maxBlocks ?? 1,
-        maxEstimatedTokens: current?.maxEstimatedTokens ?? 2048,
-        maxBlockEstimatedTokens:
-          current?.maxBlockEstimatedTokens ?? current?.maxSummaryEstimatedTokens ?? 1024,
-        highWaterName: current?.highWaterName ?? `${base.name ?? 'manual'}-manual-history-compact`,
-      },
-    };
-  }
-
-  private hasHistoryCompactWriter(): boolean {
-    return Boolean(
-      this.input.writeHistoryCompact ||
-        (this.input.summarizeHistoryCompact && this.input.recordHistoryCompactCheckpoint),
-    );
-  }
-
-  private hasHistoryCompactCheckpointWriter(): boolean {
-    return Boolean(this.input.summarizeHistoryCompact && this.input.recordHistoryCompactCheckpoint);
+    return this.compaction.compactHistory(input, this.priorRequestShape?.requestShapeHash);
   }
 
   // --------------------------------------------------------------------------
@@ -2306,7 +2173,7 @@ export class AiSdkBackend implements AgentBackend {
     }
     this.aborted = true;
     this.abortController?.abort();
-    this.historyCompactAbortController?.abort();
+    this.compaction.abortHistoryCompact();
     if (this.currentTurnId !== null) {
       this.input.permissionEngine.endTurn(this.currentTurnId, 'aborted');
       this.toolRuntime.endTurn(this.currentTurnId, 'aborted');
@@ -2401,7 +2268,8 @@ export class AiSdkBackend implements AgentBackend {
     const contextBudget = preparedContextBudget.policy;
     const budgeted = applyRuntimeEventContextBudget(priorRuntimeContext, contextBudget, {
       historyCompactProtocol:
-        contextBudget?.historyCompact?.checkpoint || this.hasHistoryCompactCheckpointWriter()
+        contextBudget?.historyCompact?.checkpoint ||
+        this.compaction.hasHistoryCompactCheckpointWriter()
           ? 'checkpoint_v2'
           : 'legacy_v1',
     });
@@ -2418,7 +2286,7 @@ export class AiSdkBackend implements AgentBackend {
     if (
       budgeted?.historyCompactBlocks?.length &&
       contextBudget?.historyCompact?.mode === 'read_write' &&
-      this.hasHistoryCompactWriter()
+      this.compaction.hasHistoryCompactWriter()
     ) {
       const loadedBlockIds = new Set(
         (contextBudget.historyCompact.blocks ?? []).map((block) => block.blockId),
